RIS.Randomizing: fix return type from ulong to uint for GetNormalizedIndex(uint) mtehod in SecureRandom class, CachedSecureRandom class, IUnbiasedRandom interface

+++ RIS/Randomizing/CachedSecureRandom.cs
@@ -562,7 +562,7 @@ namespace RIS.Randomizing
                 GetBiasZone(targetSamplingLength));
         }
 
-        public ulong GetNormalizedIndex(uint targetSamplingLength)
+        public uint GetNormalizedIndex(uint targetSamplingLength)
         {
             return (uint)(GetNormalizedUInt32(targetSamplingLength) % targetSamplingLength);
         }
