+++ RIS/Randomizing/SecureRandom.cs
@@ -306,7 +306,7 @@ namespace RIS.Randomizing
                 GetBiasZone(targetSamplingLength));
         }
 
-        public ulong GetNormalizedIndex(uint targetSamplingLength)
+        public uint GetNormalizedIndex(uint targetSamplingLength)
         {
             return (uint)(GetNormalizedUInt32(targetSamplingLength) % targetSamplingLength);
         }
